requirements.txt
biopython
matplotlib>=3.5.1
numpy>=1.21.5
pandas>=1.4.2
